refactor[#17] : shared/utils -> shared/route 슬라이스명 변경
