+++ blog/src/shared/route/createImageConfig.ts
