+++ blog/src/shared/route/readme.md
@@ -1,0 +1,5 @@
+# route
+
+해당 폴더는 라우트 핸들러에서 사용하는 메소드들을 저장하는 폴더입니다.
+
+클라이언트단과 서버단의 로직을 분리하기 위해 생성되었습니다.
